dev: reorganize files
