+++ tools/editor-tools/src/features/font-view/components/header.ts
@@ -1,12 +1,17 @@
-import van, { type State } from "vanjs-core";
-import { FONT_STRETCH_CATEGORIES, FONT_STYLE_OPTIONS, FONT_WEIGHT_CATEGORIES } from "./constants";
+import van, { type ChildDom, type State } from "vanjs-core";
+import {
+  FONT_STRETCH_CATEGORIES,
+  FONT_STYLE_CATEGORIES,
+  FONT_WEIGHT_CATEGORIES,
+} from "../constants";
+import type { FontFilterStates, FontStats } from "../filtering";
 
 const { input, div, button } = van.tags;
 
 /**
  * Creates a search input component for filtering fonts
  */
-export const SearchInput = (searchQuery: State<string>) => {
+const SearchInput = (searchQuery: State<string>) => {
   return input({
     class: "input flex",
     type: "text",
@@ -22,7 +27,7 @@ export const SearchInput = (searchQuery: State<string>) => {
 /**
  * Creates a weight filter as toggle buttons
  */
-export const WeightFilter = (weightFilter: State<string>) => {
+const WeightFilter = (weightFilter: State<string>) => {
   const selectedWeights = van.derive(() =>
     weightFilter.val ? weightFilter.val.split(",").filter(Boolean) : [],
   );
@@ -60,7 +65,7 @@ export const WeightFilter = (weightFilter: State<string>) => {
 /**
  * Creates a stretch filter as toggle buttons
  */
-export const StretchFilter = (stretchFilter: State<string>) => {
+const StretchFilter = (stretchFilter: State<string>) => {
   const selectedStretches = van.derive(() =>
     stretchFilter.val ? stretchFilter.val.split(",").filter(Boolean) : [],
   );
@@ -98,7 +103,7 @@ export const StretchFilter = (stretchFilter: State<string>) => {
 /**
  * Creates a style filter as toggle buttons
  */
-export const StyleFilter = (styleFilter: State<string>) => {
+const StyleFilter = (styleFilter: State<string>) => {
   const selectedStyles = van.derive(() =>
     styleFilter.val ? styleFilter.val.split(",").filter(Boolean) : [],
   );
@@ -117,7 +122,7 @@ export const StyleFilter = (styleFilter: State<string>) => {
     div("Style"),
     div(
       { class: "filter-options" },
-      ...Object.entries(FONT_STYLE_OPTIONS).map(([key, category]) =>
+      ...Object.entries(FONT_STYLE_CATEGORIES).map(([key, category]) =>
         button(
           {
             class: van.derive(() =>
@@ -138,7 +143,7 @@ export const StyleFilter = (styleFilter: State<string>) => {
 /**
  * Creates a clear filters button
  */
-export const ClearFiltersButton = (clearFilters: () => void) => {
+const ClearFiltersButton = (clearFilters: () => void) => {
   return button(
     {
       class: "btn",
@@ -147,5 +152,61 @@ export const ClearFiltersButton = (clearFilters: () => void) => {
       onclick: clearFilters,
     },
     "Clear Filters",
+  );
+};
+
+const ToggleButton = (
+  body: ChildDom,
+  title: string,
+  onclick: (this: HTMLButtonElement) => void,
+  active?: boolean,
+) => {
+  return button(
+    {
+      class: active ? "toggle-btn activated" : "toggle-btn",
+      title,
+      onclick,
+    },
+    body,
+  );
+};
+
+const StatsText = (stats: State<FontStats>) => {
+  const { filtered, total, variants } = stats.val;
+  const text =
+    filtered === total
+      ? `Showing ${total} font families (${variants} variants)`
+      : `Showing ${filtered} of ${total} font families (${variants} variants)`;
+  return div({ class: "font-stats" }, text);
+};
+
+export const Header = (
+  filterStates: FontFilterStates,
+  stats: State<FontStats>,
+  showNumber: State<boolean>,
+  clearFilters: () => void,
+) => {
+  return div(
+    { class: "font-view-header card" },
+    SearchInput(filterStates.searchQuery),
+    div(
+      { class: "font-filters-section" },
+      WeightFilter(filterStates.weightFilter),
+      StretchFilter(filterStates.stretchFilter),
+      StyleFilter(filterStates.styleFilter),
+    ),
+    ClearFiltersButton(clearFilters),
+    div(
+      { class: "font-stats-section" },
+      StatsText(stats),
+      ToggleButton(
+        "Show Numbers",
+        "Toggle to show weight and stretch numbers",
+        () => {
+          showNumber.val = !showNumber.val;
+        },
+        showNumber.val,
+      ),
+    ),
   );
 };
