+++ tools/editor-tools/src/features/font-view/styles.css
@@ -153,7 +153,6 @@
   align-items: center;
 }
 
-
 /* todo: redesign this */
 .font-file-info {
   font-size: 11px;
