+++ tools/editor-tools/tsconfig.json
@@ -14,6 +14,12 @@
     "isolatedModules": true,
     "noEmit": true,
 
+    /* Path mapping */
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["src/*"]
+    },
+
     /* Linting */
     "strict": true,
     "noUnusedLocals": true,
